pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["doublekai <wdoublekai@126.com>"]
[[tool.poetry.source]]
name = "tsinghua"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
pydantic = "^1.9.1"
yapl = "^0.1.5"
django-ninja = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
